requirements.txt
requests
beautifulsoup4
lxml
feedparser
PyGithub